% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitnessFunction_cpp}
\alias{fitnessFunction_cpp}
\title{Fitness function for WOA optimization (C++ implementation)}
\usage{
fitnessFunction_cpp(dist_vec, nrows, medoids)
}
\arguments{
\item{dist_vec}{Distance vector (from dist object)}

\item{nrows}{Number of rows in the original data}

\item{medoids}{Integer vector of medoid indices (1-indexed)}
}
\value{
total_dist (double), Inf if invalid clustering
}
\description{
Combines cluster assignment and total distance calculation.
Returns Inf if any cluster has fewer than 2 members.
}
\keyword{internal}
