% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitnessFunction_serial_cpp}
\alias{fitnessFunction_serial_cpp}
\title{Deterministic serial fitness function for final evaluation}
\usage{
fitnessFunction_serial_cpp(dist_vec, nrows, medoids)
}
\arguments{
\item{dist_vec}{Distance vector (from dist object)}

\item{nrows}{Number of rows in the original data}

\item{medoids}{Integer vector of medoid indices (1-indexed)}
}
\value{
total_dist (double), Inf if invalid clustering
}
\description{
Deterministic serial fitness function for final evaluation
}
\keyword{internal}
