% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{medoidsEqual_cpp}
\alias{medoidsEqual_cpp}
\title{Check if two medoid solutions are identical}
\usage{
medoidsEqual_cpp(medoids1, medoids2)
}
\arguments{
\item{medoids1}{First medoid vector (1-indexed)}

\item{medoids2}{Second medoid vector (1-indexed)}
}
\value{
TRUE if identical (regardless of order), FALSE otherwise
}
\description{
Check if two medoid solutions are identical
}
\keyword{internal}
