% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{projectToIndex_cpp}
\alias{projectToIndex_cpp}
\title{Project MDS coordinates to nearest unique sample indices}
\usage{
projectToIndex_cpp(medoid_coords, Z)
}
\arguments{
\item{medoid_coords}{NumericMatrix of medoid coordinates in MDS space (ClusNum x mds_dim)}

\item{Z}{NumericMatrix of sample coordinates in MDS space (nrows x mds_dim)}
}
\value{
IntegerVector of nearest unique sample indices (1-indexed, length = ClusNum)
}
\description{
Finds the nearest sample in the MDS embedded space for each medoid coordinate,
while enforcing a one-to-one mapping between medoids and samples (no duplicates).
}
\keyword{internal}
